style: Adjust layout of home page
