+++ src/pages/index.module.css
@@ -7,13 +7,6 @@
     width: 50vw;
     height: 55vh;
 }
-.descriptionContainer{
-    display: flex;
-    width: 50vw;
-    border: 1px solid grey;
-    height: 55vh;
-    justify-content: center;
-}
 .bottomCraphContainer{
     width: 100vw;
     height: 36.5vh;
@@ -30,20 +23,22 @@
     padding: 1.25rem;
     width: 50vw;
 }
+.beachInfoContainer {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    gap: 5rem/* 80px */;
+}
 
-@media (max-width: 1600px) {
+@media (max-width: 1400px) {
     .flexWrap1000{
         flex-wrap: wrap;
     }
     .mapContainer{
         width: 100vw;
         height: 50vh;
-    }
-    .descriptionContainer{
-        justify-content: center;
-        align-items: center;
-        width: 100vw;
-        height: 55vh;
     }
     .bottomCraphContainer{
         height: 100vh;
@@ -59,12 +54,9 @@
     }
 }
 
-@media (max-width: 800px) {
-    .descriptionContainer{
-        flex-wrap: wrap;
-        height: 100vh;
-        justify-content: center;
-        align-items: center;
+@media (max-width: 700px) {
+    .beachInfoContainer{
+        gap: 1rem;
     }
 }
